refactor(capabilities): :recycle: migrate header organisms to component-store pattern
BrandHeader: pure component (only useGlassmorphism, allowlisted). Preserves
wrapper?:FC<PropsWithChildren> inversion prop — no features/ import.
GenericHeader: ComponentStore<{isSubtitleExpanded},{toggleSubtitle}> via
useGenericHeaderStore. Both moved to header/brand-header/ and
header/generic-header/ sub-folders with index.ts barrels.
Old flat header/brand-header.tsx and header/generic-header.tsx removed.
