+++ src/components/design-system/organism/header/brand-header/index.ts
@@ -1,0 +1,1 @@
+export { BrandHeader } from "./brand-header";
