+++ src/components/design-system/organism/header/generic-header/index.ts
@@ -1,0 +1,2 @@
+export { GenericHeader } from "./generic-header";
+export type { ChatHeaderProps } from "./generic-header";
